Python JDBC script for reading IPEDS Access Database
I want to remove Jason Bryer's R package jbyer/ipeds from my data
pipeline and read the data directly from the .accdb files published by
NCES as part of the anual IPEDS report. My dev environment runs MacOs
and Microsoft does not supply drivers for Access Databases to
non-Windows platforms. I am using the JDBC driver UCanAccess along with
the Python library JayDeBeApi in order to interface with the Access DB.
Rlang has a package called RJDBC which allows for using JDBC in R but
the support documentation was not forthcoming and did not seem like a a
good choice. The basic script for interfacing with the accdb files is
written but needs a good practices pass and includes some test queries
that will need to be refactored for the next iteration. I plan to still
save the output to .Rda but I also plan to offer a language agnostic
option of .csv. This project may be forked into the Rlang specific
branch and the Python based csv branch in future.